Major improvements to session and tunnel modularity
The tunnel is the main session protocol. The core features are
opening new stream connections, closing the tunnel, and forwarding
other messages to an encapsulated sub-protocol.

Sub-protocols can be created and plugged into a tunnel. The main
features needed for a sub-protocol inlcude:

- The SessionBuilder trait must be implemented to define how to
  create new sub-protocol sessions and split them into read/write
  halves.

- The Sink trait must (still) be implemented to handle incoming
  messages of type TunnelMessage::Encapsulated. The sub-protocol
  can define its own Codec to decode encapsulated bytes as needed.

- The Stream trait must (still) be implemented to provide new
  messages of type TunnelMessage::Encapsulated. The sub-protocol
  can define its own message type as long as it can be encoded
  into bytes that are carried in a TunnelMessage.

Added a new BytesSession as the simplest possible sub-protocol.
The BytesSession is an adapter that just forwards raw bytes back
and forth between inner AsyncRead and AsyncWrite objects and the
outer Tunnel protocol.

Improved tunnel tests and configured the pt and socks mode to use
the BytesSession for now. TurboSession was removed for now, but
will be added back in a future commit following the sub-protocol
framework.